Much of the design is in place and it builds now.
It doesn't output anything onto the e-paper display yet,
and hasn't been tested at all, so don't use it on a real
device yet because it may permanently break your device
at the moment!

A big thanks is owed to the author of this git repository
for giving me some ideas of where to start, to get
something working:
https://github.com/martinberlin/cale-idf

And a thanks is owed to the author of this git repository
for refusing to fix their code or take any of my fixes,
which ultimately inspired me to start making this
project:
https://github.com/ZinggJM/GxEPD

Expect some working examples soonish.